+++ typescript/src/adapters/openai/backend/client.ts
@@ -12,24 +12,12 @@ export type OpenAIClientSettings = OpenAIProviderSettings;
 
 export class OpenAIClient extends BackendClient<OpenAIClientSettings, OpenAIProvider> {
   protected create(): OpenAIProvider {
-    const extraHeaders = parseHeadersFromEnv("OPENAI_API_HEADERS");
-
-    const baseURL = this.settings?.baseURL || getEnv("OPENAI_API_ENDPOINT");
-    let compatibility: string | undefined =
-      this.settings?.compatibility || getEnv("OPENAI_COMPATIBILITY_MODE");
-    if (baseURL && !compatibility) {
-      compatibility = "compatible";
-    } else if (!baseURL && !compatibility) {
-      compatibility = "strict";
-    }
-
     return createOpenAI({
       ...this.settings,
-      compatibility: compatibility as "strict" | "compatible" | undefined,
       apiKey: this.settings?.apiKey || getEnv("OPENAI_API_KEY"),
-      baseURL,
+      baseURL: this.settings?.baseURL || getEnv("OPENAI_API_ENDPOINT"),
       headers: {
-        ...extraHeaders,
+        ...parseHeadersFromEnv("OPENAI_API_HEADERS"),
         ...this.settings?.headers,
       },
       fetch: vercelFetcher(this.settings?.fetch),